[mega wip] start on search/explore page
summary of explore page requirements:
* all filters optional, including search query
* control over result ordering
* page size of 20ish results?
* ideally reuse the "cards grouped together / cards grouped by source" UI from grid selector
