+++ frontend/src/features/gridSelector/GridSelectorModal.tsx
@@ -6,8 +6,6 @@
 
 import React, {
   FormEvent,
-  PropsWithChildren,
-  ReactElement,
   useCallback,
   useMemo,
   useRef,
@@ -15,59 +13,18 @@ import React, {
 } from "react";
 import Button from "react-bootstrap/Button";
 import Col from "react-bootstrap/Col";
-import Collapse from "react-bootstrap/Collapse";
 import Form from "react-bootstrap/Form";
 import Modal from "react-bootstrap/Modal";
 import Row from "react-bootstrap/Row";
-import Stack from "react-bootstrap/Stack";
-// @ts-ignore: https://github.com/arnthor3/react-bootstrap-toggle/issues/21
-import Toggle from "react-bootstrap-toggle";
-import RenderIfVisible from "react-render-if-visible";
 
-import { ToggleButtonHeight } from "@/common/constants";
 import { CardDocument, useAppDispatch, useAppSelector } from "@/common/types";
-import { RightPaddedIcon } from "@/components/icon";
-import { MemoizedEditorCard } from "@/features/card/Card";
+import { AutofillCollapse } from "@/components/AutofillCollapse";
+import { CardResultSet } from "@/features/card/CardResultSet";
 import { selectCardDocumentsByIdentifier } from "@/store/slices/cardDocumentsSlice";
-import { selectSourceNamesByKey } from "@/store/slices/sourceDocumentsSlice";
 import {
-  makeAllSourcesInvisible,
-  makeAllSourcesVisible,
-  selectAnySourcesCollapsed,
-  selectFacetBySource,
   selectJumpToVersionVisible,
-  selectSourcesVisible,
-  toggleFacetBySource,
   toggleJumpToVersionVisible,
-  toggleSourceVisible,
 } from "@/store/slices/viewSettingsSlice";
-
-interface CardGridCardProps {
-  identifier: string;
-  index: number;
-  selectImage: {
-    (identifier: string): void;
-  };
-  selectedImage?: string;
-}
-
-function CardGridCard({
-  identifier,
-  index,
-  selectImage,
-  selectedImage,
-}: CardGridCardProps) {
-  return (
-    <MemoizedEditorCard
-      imageIdentifier={identifier}
-      cardHeaderTitle={`Option ${index + 1}`}
-      cardOnClick={() => selectImage(identifier)}
-      key={`gridSelector-${identifier}`}
-      noResultsFound={false}
-      highlight={identifier === selectedImage}
-    />
-  );
-}
 
 interface GridSelectorProps {
   title?: string;
@@ -84,192 +41,6 @@ interface GridSelectorProps {
   };
 }
 
-interface CardGridDisplayProps {
-  cardIdentifiersAndOptionNumbersBySource: {
-    [sourceKey: string]: Array<[string, number]>;
-  };
-  selectImage: {
-    (identifier: string): void;
-  };
-  selectedImage?: string;
-  sourceNamesByKey: { [sourceKey: string]: string };
-}
-
-interface AutofillCollapseProps {
-  expanded: boolean;
-  onClick: () => void;
-  zIndex: number;
-  title: ReactElement | string;
-  subtitle?: string;
-  children: ReactElement;
-  sticky?: boolean;
-}
-
-/**
- * bit of a shitty component name sorry
- * @param children Children to render in the body of this collapsible man
- * @param expanded Whether this collapsible man is expanded
- * @param onClick What to do when trying to expand this man
- * @param zIndex The base z-index of this man
- * @param title The title of this man
- * @param subtitle Optionally, the subtitle of this man
- * @param sticky Whether or not the man's collapse bar is sticky
- * @constructor
- */
-function AutofillCollapse({
-  children,
-  expanded,
-  onClick,
-  zIndex,
-  title,
-  subtitle,
-  sticky = false,
-}: AutofillCollapseProps) {
-  return (
-    <>
-      <div
-        className={sticky ? "sticky-top" : ""}
-        onClick={onClick}
-        style={{
-          backgroundColor: "#4E5D6B",
-          zIndex: zIndex + 100,
-        }}
-      >
-        <hr className="mt-0" />
-        <Stack direction="horizontal" gap={2} className="d-flex ps-3 pe-3">
-          {title}
-          {subtitle && (
-            <h6 className="text-primary prevent-select">{subtitle}</h6>
-          )}
-
-          <button className="ms-auto bg-transparent border-0">
-            <h4
-              className={`bi bi-chevron-left rotate-${expanded ? "" : "neg"}90`}
-              style={{ transition: "all 0.25s 0s" }}
-            />
-          </button>
-        </Stack>
-        <hr className="mb-0" />
-      </div>
-      <div className="py-2" />
-      <Collapse in={expanded}>
-        <div>{children}</div>
-      </Collapse>
-    </>
-  );
-}
-
-function CardRow({ children }: PropsWithChildren) {
-  return (
-    <Row className="g-0 px-3" xxl={6} xl={6} lg={4} md={3} sm={2} xs={2}>
-      {children}
-    </Row>
-  );
-}
-
-/**
- * Render all images in `cardIdentifiersAndOptionNumbersBySource` in one block -
- * i.e. not separated by source.
- */
-function CardsGroupedTogether({
-  cardIdentifiersAndOptionNumbersBySource,
-  selectImage,
-  selectedImage,
-  sourceNamesByKey,
-}: CardGridDisplayProps) {
-  return (
-    <CardRow>
-      {Object.entries(cardIdentifiersAndOptionNumbersBySource).map(
-        ([key, value], sourceIndex) => (
-          <>
-            {value.flatMap(([identifier, index]) => (
-              <RenderIfVisible
-                key={`gridSelector-${identifier}-wrapper`}
-                initialVisible={index < 20}
-                visibleOffset={500}
-                stayRendered
-              >
-                <CardGridCard
-                  identifier={identifier}
-                  index={index}
-                  selectImage={selectImage}
-                  selectedImage={selectedImage}
-                />
-              </RenderIfVisible>
-            ))}
-          </>
-        )
-      )}
-    </CardRow>
-  );
-}
-
-/**
- * Render all images in `cardIdentifiersAndOptionNumbersBySource` separated by source.
- * Allow users to toggle whether each source's cards are showed/hidden.
- */
-function CardsFacetedBySource({
-  cardIdentifiersAndOptionNumbersBySource,
-  selectImage,
-  selectedImage,
-  sourceNamesByKey,
-}: CardGridDisplayProps) {
-  //# region queries and hooks
-
-  const dispatch = useAppDispatch();
-  const sourcesVisible = useAppSelector(selectSourcesVisible);
-
-  //# endregion
-
-  return (
-    <>
-      {Object.entries(cardIdentifiersAndOptionNumbersBySource).map(
-        ([sourceKey, cardIdentifiersAndOptionNumbers], sourceIndex) => (
-          <>
-            <AutofillCollapse
-              key={sourceKey}
-              expanded={sourcesVisible[sourceKey] ?? true}
-              onClick={() => dispatch(toggleSourceVisible(sourceKey))}
-              zIndex={sourceIndex}
-              title={
-                <h3
-                  className="orpheus prevent-select"
-                  style={{ fontStyle: "italic" }}
-                >
-                  {sourceNamesByKey[sourceKey]}
-                </h3>
-              }
-              subtitle={`${cardIdentifiersAndOptionNumbers.length} version${
-                cardIdentifiersAndOptionNumbers.length != 1 ? "s" : ""
-              }`}
-              sticky
-            >
-              <CardRow key={`${sourceKey}-row`}>
-                {cardIdentifiersAndOptionNumbers.map(
-                  ([identifier, optionNumber]) => (
-                    <RenderIfVisible
-                      key={`gridSelector-${identifier}-wrapper`}
-                      initialVisible={optionNumber < 20}
-                    >
-                      <CardGridCard
-                        identifier={identifier}
-                        index={optionNumber}
-                        selectImage={selectImage}
-                        selectedImage={selectedImage}
-                      />
-                    </RenderIfVisible>
-                  )
-                )}
-              </CardRow>
-            </AutofillCollapse>
-            <div className="py-2" />
-          </>
-        )
-      )}
-    </>
-  );
-}
-
 export function GridSelectorModal({
   title = "Select Version",
   testId,
@@ -282,12 +53,7 @@ export function GridSelectorModal({
   //# region queries and hooks
 
   const dispatch = useAppDispatch();
-  const cardDocuments = useAppSelector((state) =>
-    selectCardDocumentsByIdentifier(state, imageIdentifiers)
-  );
-  const facetBySource = useAppSelector(selectFacetBySource);
-  const sourceNamesByKey = useAppSelector(selectSourceNamesByKey);
-  const anySourcesCollapsed = useAppSelector(selectAnySourcesCollapsed);
+
   const jumpToVersionVisible = useAppSelector(selectJumpToVersionVisible);
 
   //# endregion
@@ -326,33 +92,6 @@ export function GridSelectorModal({
     ((optionNumber ?? 0) > 0 &&
       (optionNumber ?? 0) < imageIdentifiers.length + 1) ||
     (imageIdentifier !== "" && imageIdentifiers.includes(imageIdentifier));
-  const sourceKeys = Object.keys(sourceNamesByKey);
-  const cardIdentifiersAndOptionNumbersBySource = useMemo(
-    () =>
-      imageIdentifiers.reduce(
-        (
-          accumulator: { [sourceKey: string]: Array<[string, number]> },
-          value,
-          currentIndex
-        ) => {
-          const cardDocument: CardDocument | null = cardDocuments[value];
-          if (cardDocument != null) {
-            if (
-              !Object.prototype.hasOwnProperty.call(
-                accumulator,
-                cardDocument.source
-              )
-            ) {
-              accumulator[cardDocument.source] = [];
-            }
-            accumulator[cardDocument.source].push([value, currentIndex]);
-          }
-          return accumulator;
-        },
-        {}
-      ),
-    [cardDocuments, imageIdentifiers]
-  );
 
   //# endregion
 
@@ -432,70 +171,12 @@ export function GridSelectorModal({
             <hr />
           </>
         </AutofillCollapse>
-        <div className="px-3 pb-3">
-          <Row>
-            <h4>Browse Versions</h4>
-            <Col md={8} sm={6}>
-              <Stack direction="horizontal" gap={2}>
-                <span className="me-auto">Show&nbsp;All&nbsp;Cards...</span>
-                <Toggle
-                  onClick={() => dispatch(toggleFacetBySource())}
-                  on="Grouped By Source"
-                  onClassName="flex-centre"
-                  off="Grouped Together"
-                  offClassName="flex-centre"
-                  onstyle="success"
-                  offstyle="info"
-                  width={100 + "%"}
-                  size="md"
-                  height={ToggleButtonHeight + "px"}
-                  active={facetBySource}
-                />
-              </Stack>
-            </Col>
-            {facetBySource && (
-              <Col md={4} sm={6}>
-                <div className="d-grid g-0">
-                  <Button
-                    onClick={() =>
-                      dispatch(
-                        anySourcesCollapsed
-                          ? makeAllSourcesVisible()
-                          : makeAllSourcesInvisible(sourceKeys)
-                      )
-                    }
-                  >
-                    <RightPaddedIcon
-                      bootstrapIconName={`arrows-${
-                        anySourcesCollapsed ? "expand" : "collapse"
-                      }`}
-                    />{" "}
-                    {anySourcesCollapsed ? "Expand" : "Collapse"} All
-                  </Button>
-                </div>
-              </Col>
-            )}
-          </Row>
-        </div>
-        {facetBySource ? (
-          <CardsFacetedBySource
-            cardIdentifiersAndOptionNumbersBySource={
-              cardIdentifiersAndOptionNumbersBySource
-            }
-            selectImage={selectImage}
-            selectedImage={selectedImage}
-            sourceNamesByKey={sourceNamesByKey}
-          />
-        ) : (
-          <CardsGroupedTogether
-            cardIdentifiersAndOptionNumbersBySource={
-              cardIdentifiersAndOptionNumbersBySource
-            }
-            selectImage={selectImage}
-            selectedImage={selectedImage}
-            sourceNamesByKey={sourceNamesByKey}
-          />
-        )}
+        <CardResultSet
+          headerText="Select Version"
+          imageIdentifiers={imageIdentifiers}
+          handleClick={selectImage}
+          selectedImage={selectedImage}
+        />
       </Modal.Body>
       <Modal.Footer>
         <Button variant="secondary" onClick={handleClose}>
